xray_analyzer/requirements.txt
Flask
flask-cors
torch
torchvision
numpy
opencv-python
Pillow
